feat: apply Real Book theme
