+++ src/style.css
@@ -1,14 +1,11 @@
 :root {
   --bg: #fff;
   --text: #000;
-  --header-bg: #f0f0f0;
-  --header-text: #000;
-  --rail-bg: #ddd;
-  --measure-border: #888;
-  --secondary-color: #555;
+  --measure-border: #000;
+  --secondary-color: #000;
 }
 
-/* reset suave (no toca tu lógica ni componentes) */
+/* reset */
 *,
 *::before,
 *::after { box-sizing: border-box; }
@@ -19,78 +16,77 @@ body {
   line-height: 1.35;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background: var(--bg, #fff);
-  color: var(--text, #000);
+  background: var(--bg);
+  color: var(--text);
   font-size: var(--font-size, 16px);
 }
 #app { padding: 12px; max-width: 1200px; margin: 0 auto; }
 button, input, select { font: inherit; }
 button { cursor: pointer; }
-input[type="number"], input[type="text"], select {
-  padding: 4px 6px; border: 1px solid #c9c9c9; border-radius: 4px;
-}
+input[type="number"], input[type="text"], select { padding: 4px 6px; border: 1px solid #c9c9c9; border-radius: 4px; }
 label { display: inline-block; margin-right: 6px; }
 
 body.dark {
-  --bg: #1e1e1e;
+  --bg: #000;
   --text: #fff;
-  --header-bg: #222;
-  --header-text: #fff;
-  --rail-bg: #444;
-  --measure-border: #666;
-  --secondary-color: #bbb;
+  --measure-border: #fff;
+  --secondary-color: #ccc;
 }
 
 header {
-  background: var(--header-bg, #222);
-  color: var(--header-text, #fff);
-  padding: 0.5rem 1rem;
+  background: var(--bg);
+  color: var(--text);
+  text-align: center;
+  padding: 1rem 0;
+}
+.chart-title {
+  margin: 0;
+  font-size: 2rem;
+  font-weight: bold;
 }
 
-.net-status {
-  margin-left: 1rem;
-  font-size: 0.9rem;
-}
-
-.rail {
-  height: 20px;
-  background: var(--rail-bg, #444);
-}
+.rail { display: none; }
 
 .grid {
   display: flex;
   flex-direction: column;
-  padding: 1rem;
+  padding: 1rem 0;
 }
 
 .section {
-  margin-bottom: 1rem;
   display: flex;
+  align-items: flex-start;
+  margin-bottom: 1rem;
   position: relative;
 }
 
 .section-title {
+  background: #000;
+  color: #fff;
+  width: 1.5rem;
+  height: 1.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-right: 0.5rem;
   font-weight: bold;
-  margin-bottom: 0.5rem;
+}
+body.dark .section-title {
+  background: #ccc;
+  color: #000;
 }
 
 .measure {
-  border: 1px solid var(--measure-border, #888);
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
+  display: flex;
   width: 8rem;
   height: 4rem;
-  margin-right: 0.25rem;
+  margin-right: 1rem;
   position: relative;
+  border-left: 2px solid var(--measure-border);
+  border-right: 2px solid var(--measure-border);
 }
-
-.measure.drag-over {
-  border-color: #00f;
-}
-
-.measure.selected {
-  border-color: #f00;
-}
+.measure.drag-over { border-left-color: #00f; border-right-color: #00f; }
+.measure.selected { border-left-color: #f00; border-right-color: #f00; }
 
 .markers {
   position: absolute;
@@ -98,7 +94,7 @@ header {
   left: 0;
   width: 100%;
   text-align: center;
-  font-size: 0.75rem;
+  font-size: 0.9rem;
 }
 
 .volta-container {
@@ -111,14 +107,14 @@ header {
 .volta {
   position: absolute;
   top: -1.5rem;
-  border: 1px solid #000;
+  border: 1px solid var(--text);
   border-bottom: none;
   font-size: 0.75rem;
   padding: 0 0.25rem;
 }
 
 .slot {
-  border-right: 1px solid var(--measure-border, #888);
+  flex: 1;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -126,21 +122,28 @@ header {
 }
 
 .chord {
-  line-height: 1.2;
+  font-size: 1.8rem;
+  font-weight: bold;
+  line-height: 1;
+  color: var(--text);
 }
 
 .secondary {
-  font-size: 0.7em;
+  font-size: 0.6em;
   line-height: 1;
-  color: var(--secondary-color, #555);
-}
-
-.slot:last-child {
-  border-right: none;
+  color: var(--secondary-color);
 }
 
 .controls {
-  margin: 1rem;
+  margin: 1rem 0;
+  color: #666;
+  font-size: 0.8rem;
+}
+.controls button,
+.controls input,
+.controls select,
+.controls label {
+  font-size: inherit;
 }
 
 .message {
@@ -160,20 +163,23 @@ header {
 }
 
 @media (max-width: 600px) {
-  .grid {
-    padding: 0.5rem;
-  }
-
-  .measure {
-    width: 100%;
-    height: 3rem;
-  }
+  .grid { padding: 0.5rem 0; }
+  .measure { width: 100%; height: 3rem; }
 }
 
 @media print {
-  .secondary {
+  body {
+    --bg: #fff;
+    --text: #000;
+    --measure-border: #000;
+    --secondary-color: #000;
+    background: #fff;
     color: #000;
   }
+  .controls,
+  .rail { display: none; }
+  .chord { font-size: 2rem; color: #000; }
+  .secondary { color: #000; }
 }
 
 .library-modal {
@@ -190,8 +196,8 @@ header {
 }
 
 .library-modal-content {
-  background: var(--bg, #fff);
-  color: var(--text, #000);
+  background: var(--bg);
+  color: var(--text);
   padding: 1rem;
   max-height: 80vh;
   overflow: auto;
@@ -212,4 +218,3 @@ header {
 .library-modal li {
   margin: 0.25rem 0;
 }
-
